pyproject.toml
[tool.poetry]
name = "programatically-create-jira-issues-from-github"
version = "0.1.0"
description = ""
authors = ["githubofkrishnadhas <krishnadhasnk1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
atlassian-python-api = "^3.41.13"
python-dotenv = "^1.0.1"
requests = "^2.32.2"
argparse = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
